+++ lib/src/l10n/_l10n.dart
@@ -11,5 +11,6 @@ export './languages/croppy_es.dart';
 export './languages/croppy_zh.dart';
 export './languages/croppy_tr.dart';
 export './languages/croppy_pl.dart';
+export './languages/croppy_ro.dart';
 export './croppy_localizations.dart';
 export './l10n_provider.dart';